+++ src/Position.h
@@ -1,14 +1,14 @@
 #ifndef POSITION_H
 #define POSITION_H
 class Position{
-    int x;
-    int y;
+    float x;
+    float y;
 public:
-    Position(int x, int y);
-    int getX();
-    void setX(int x);
-    int getY();
-    void setY(int y);
+    Position(float x, float y);
+    float getX();
+    void setX(float x);
+    float getY();
+    void setY(float y);
 
 };
 #endif
